PI/CogniKids/cognikids-backend/requirements.txt
Flask
python-dotenv
pymongo
Flask-PyMongo
bcrypt
Flask-Bcrypt
Flask-Cors
PyJWT
flasgger
flask-swagger-ui
pydantic
requests
streamlit
redis
paho-mqtt
scikit-learn
pandas
joblib
pytz